Change the button disable state checker. (#54)
Create a new button state checker function in the surface module.
This function will check the target type first. If the target is an i
element then it will check the element parent node, otherwise it will test
the target itself. This was necessary to avoid the warnings when
the user clicks to an i element (which is inside the button as an icon).

IoT.jsCode-DCO-1.0-Signed-off-by: Imre Kiss kissi.szeged@partner.samsung.com